requirements.txt
streamlit
pandas
numpy
openai
matplotlib
plotly
scikit-learn
python-dotenv
requests